requirements.txt
selenium==4.15.0
pytest==7.4.0
allure-pytest==2.13.2
webdriver-manager==4.0.1
pytest-html==4.0.0
pytest-metadata==3.1.1